requirements.txt
# Core Flask and database
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

google-genai==1.21.1

# Vector database and embeddings
sentence-transformers>=4.1.0
chromadb==0.5.0
huggingface-hub>=0.33.0
numpy<2.0.0

# PDF processing
PyPDF2==3.0.1

# Core utilities
pydantic==2.5.0
gunicorn==21.2.0
boto3==1.34.0
requests==2.32.3 
zappa==0.59.0

